+++ Bugsnag/Configuration/BugsnagConfiguration.m
@@ -77,6 +77,7 @@ static NSUserDefaults *userDefaults;
     // Omit apiKey - it's set explicitly in the line above
 #if BSG_HAVE_APP_HANG_DETECTION
     [copy setAppHangThresholdMillis:self.appHangThresholdMillis];
+    [copy setReportBackgroundAppHangs:self.reportBackgroundAppHangs];
 #endif
     [copy setAppType:self.appType];
     [copy setAppVersion:self.appVersion];